Simplify ROM into a const block, increase buffer size

+++ src/blaster/port.rs
@@ -1,3 +1,4 @@
+use core::ops::Index;
 use hal::prelude::*;
 
 pub struct Port {
@@ -133,8 +134,8 @@ impl Port {
     }
 
     pub fn handle<
-        A1: arrayvec::Array<Item = u8, Index = u8>,
-        A2: arrayvec::Array<Item = u8, Index = u8>,
+        A1: arrayvec::Array<Item = u8, Index = u16>,
+        A2: arrayvec::Array<Item = u8, Index = u16>,
     >(
         &mut self,
         receive: &mut arrayvec::ArrayVec<A1>,
